Move files that are in fact shared, into the Shared directory
