+++ Shared/ConfigDefaults.h
@@ -2,6 +2,9 @@
 #ifndef CONFIG_DEFAULTS_H_INCLUDED
 #define CONFIG_DEFAULTS_H_INCLUDED
 
+// 64mm in meters
+#define IPD_DEFAULT 0.064f
+
 #define DEFAULT_YAW_MULTIPLIER 25.0f
 #define DEFAULT_PITCH_MULTIPLIER 25.0f
